renamed add member
